requirements.txt
rich
requests
pyyaml
